refactor(tui): split application state handlers
